crypto/kyber: fix UBSAN signed-overflow in montgomery_reduce across variants
Problem
- UBSAN reported runtime UB: signed integer overflow in Kyber montgomery_reduce() on expression 'u = a * QINV'.
- Repro surfaced in dap_enc_kyber512_key_generate() call chain (ntt -> montgomery_reduce).
- In C, this multiplication is evaluated as signed int arithmetic before truncation to int16_t, so overflow is undefined behavior.

What changed
- Updated all Kyber reduce implementations that used the pattern 'u = a * QINV;' to:
  'u = (int16_t)a * QINV;'
- This change was applied consistently to clean/reference/optimized trees for:
  kyber512, kyber768, kyber1024 and 90s variants.
- No public headers, exported symbols, or function signatures were changed.

Why this is correct
- Montgomery reduction in this code path only needs the low 16 bits contribution for u.
- Casting a to int16_t before multiply keeps signed-int multiplication within defined range and preserves intended modulo-2^16 behavior of the original algorithm.
- The resulting arithmetic contract of montgomery_reduce() stays the same, but UB is removed under UBSAN.

Validation
- Before patch: targeted mini-repro for each affected reduce.c produced UBSAN signed-overflow reports.
- After patch: same mini-repros are clean (no UBSAN overflow) for all affected files.
- SDK-level repro with DAP_UBSAN=1 for dap_enc_kyber512_key_generate() now runs without UBSAN runtime errors.

Compatibility
- API/ABI preserved.
- Change is internal to Kyber arithmetic implementation.

(cherry picked from commit 4c67be5ee44624cbb7dc4e3ce9f8963df4cd6d03)

+++ module/crypto/src/Kyber/crypto_kem/kyber1024-90s/clean/reduce.c
@@ -18,7 +18,7 @@ int16_t montgomery_reduce(int32_t a) {
     int32_t t;
     int16_t u;
 
-    u = a * QINV;
+    u = (int16_t)a * QINV;
     t = (int32_t)u * KYBER_Q;
     t = a - t;
     t >>= 16;
